Module nemo::peano

Peano numbers are natural numbers expressed as successions of zero. We use them in our API to provide "nested" protocol escaping, as you must indicate the number of layers you wish to escape from.

Structs

S
Peano numbers: Increment

Z
Peano numbers: Zero

Traits

Peano
Represents a peano number.

Pop
This represents the types obtained by popping N layers from a stack.